Refactoring and cleaning.
